Refactor ThirdPartyAuth functions.
